Actaulización readme y uso de .bat
