+++ main.py
@@ -19,15 +19,18 @@ Buenas prácticas:
 
 
 import os
-from clean_and_backup_data import clean_and_backup_data
-from export_sql_to_excel import export_query_to_excel
-from export_estado_analysis import export_estado_analysis_xlsx
+from scripts.clean_and_backup_data import clean_and_backup_data
+from scripts.export_sql_to_excel import export_query_to_excel
+from scripts.export_estado_analysis import export_estado_analysis_xlsx
 from scripts.clean_muestra import clean_muestra
 from scripts.clean_estados import clean_estados
 from scripts.obtain_data import obtain_data
 from scripts.test_db_connection import probar_conexion_db
 from scripts.load_to_sql import main as load_to_sql_main
 from dotenv import load_dotenv
+
+import subprocess
+import sys
 import pandas as pd
 from sqlalchemy import create_engine
 
@@ -146,8 +149,25 @@ def run_pipeline():
     else:
         print("Ejecución finalizada. No se cargaron los datos a la base de datos.")
 
+
+# =======================
+# Función para análisis exploratorio
+# =======================
+def preguntar_analisis_exploratorio():
+    """
+    Pregunta al usuario si desea realizar un análisis exploratorio de los datos
+    y ejecuta main_analysis.py si la respuesta es afirmativa.
+    """
+    print("\n¿Desea realizar un análisis exploratorio de los datos?")
+    respuesta = input('Escriba "Si" para ejecutar el análisis, o cualquier otra tecla para finalizar: ').strip().lower()
+    if respuesta == 'si':
+        subprocess.run([sys.executable, 'main_analysis.py'])
+    else:
+        print("Ejecución finalizada.")
+
 # =======================
 # Punto de entrada del script principal
 # =======================
 if __name__ == '__main__':
     run_pipeline()
+    preguntar_analisis_exploratorio()
